requirements.txt
pyserial
PyYAML
InquirerPy
